Feature: Add front end for topic events view based on offset range (#2614)
* Feature: Add front end for topic events view based on offset range


* minor code quality changes


* ran mvn spotless apply


* minor html change


* coral changes with tests


* removed changes that are not needed


* fixed custom and range stale value issue


* removed all changes from topic-type.ts


* partition id validations in coral and legacy


---------

+++ coral/src/app/features/topics/details/messages/useMessagesFilters.tsx
@@ -4,22 +4,29 @@ import { useSearchParams } from "react-router-dom";
 interface FilterErrors {
   partitionIdFilters: string | null;
   customOffsetFilters: string | null;
+  rangeOffsetStartFilters: string | null;
+  rangeOffsetEndFilters: string | null;
 }
 
-const defaultOffsets = ["5", "25", "50", "custom"] as const;
+const defaultOffsets = ["5", "25", "50", "custom", "range"] as const;
 type DefaultOffset = (typeof defaultOffsets)[number];
+
+const fetchModeTypes = ["default", "custom", "range"] as const;
+type TopicMessagesFetchModeTypes = (typeof fetchModeTypes)[number];
 
 const NAMES = {
   defaultOffset: "defaultOffset",
   customOffset: "customOffset",
+  rangeOffsetStart: "rangeOffsetStart",
+  rangeOffsetEnd: "rangeOffsetEnd",
   partitionId: "partitionId",
 };
 const initialDefaultOffset: (typeof defaultOffsets)[0] = "5";
 
 interface OffsetFilters {
-  validateFilters: () => boolean;
+  validateFilters: (totalNumberOfPartitions: number) => boolean;
   filterErrors: FilterErrors;
-  getFetchingMode: () => "Custom" | "Default";
+  getFetchingMode: () => TopicMessagesFetchModeTypes;
   defaultOffsetFilters: {
     defaultOffset: DefaultOffset;
     setDefaultOffset: (defaultOffset: DefaultOffset) => void;
@@ -29,6 +36,14 @@ interface OffsetFilters {
     customOffset: string | null;
     setCustomOffset: (customOffset: string) => void;
     deleteCustomOffset: () => void;
+  };
+  rangeOffsetFilters: {
+    rangeOffsetStart: string | null;
+    setRangeOffsetStart: (rangeOffsetStart: string) => void;
+    deleteRangeOffsetStart: () => void;
+    rangeOffsetEnd: string | null;
+    setRangeOffsetEnd: (rangeOffsetEnd: string) => void;
+    deleteRangeOffsetEnd: () => void;
   };
   partitionIdFilters: {
     partitionId: string | null;
@@ -47,46 +62,82 @@ function isDefaultOffset(
 
 function useMessagesFilters(): OffsetFilters {
   const [searchParams, setSearchParams] = useSearchParams();
-  const defaultOffset = searchParams.get(NAMES.defaultOffset);
 
   const [filterErrors, setFilterErrors] = useState<FilterErrors>({
     partitionIdFilters: null,
     customOffsetFilters: null,
+    rangeOffsetStartFilters: null,
+    rangeOffsetEndFilters: null,
   });
 
-  function validateFilters() {
-    if (getFetchingMode() === "Default") {
+  function validateFilters(totalNumberOfPartitions: number) {
+    if (getFetchingMode() === "default") {
       return true;
     }
 
     const partitionIdFiltersError =
       getPartitionId() === "" || getPartitionId() === null
         ? "Please enter a partition ID"
-        : null;
-    const customOffsetFiltersError =
-      getCustomOffset() === "" || getCustomOffset() === null
-        ? "Please enter the number of recent offsets you want to view"
-        : Number(getCustomOffset()) > 100
-          ? "Entered value exceeds the view limit for offsets: 100"
-          : null;
+        : Number(getPartitionId()) < 0
+          ? "Partition ID cannot be negative"
+          : Number(getPartitionId()) >= totalNumberOfPartitions
+            ? "Invalid partition ID"
+            : null;
+
+    let customOffsetFiltersError = null;
+    let rangeOffsetStartFiltersError = null;
+    let rangeOffsetEndFiltersError = null;
+    if (getFetchingMode() === "custom") {
+      customOffsetFiltersError =
+        getCustomOffset() === "" || getCustomOffset() === null
+          ? "Please enter the number of recent offsets you want to view"
+          : Number(getCustomOffset()) > 100
+            ? "Entered value exceeds the view limit for offsets: 100"
+            : null;
+    } else {
+      rangeOffsetStartFiltersError =
+        getRangeOffsetStart() === "" || getRangeOffsetStart() === null
+          ? "Please enter the start offset"
+          : Number(getRangeOffsetStart()) < 0
+            ? "Start offset cannot be negative."
+            : null;
+      rangeOffsetEndFiltersError =
+        getRangeOffsetEnd() === "" || getRangeOffsetEnd() === null
+          ? "Please enter the end offset"
+          : Number(getRangeOffsetEnd()) < 0
+            ? "End offset cannot be negative."
+            : null;
+
+      if (
+        rangeOffsetStartFiltersError === null &&
+        rangeOffsetEndFiltersError === null &&
+        Number(getRangeOffsetStart()) > Number(getRangeOffsetEnd())
+      ) {
+        rangeOffsetStartFiltersError = "Start must me less than end.";
+      }
+    }
 
     setFilterErrors({
       partitionIdFilters: partitionIdFiltersError,
       customOffsetFilters: customOffsetFiltersError,
+      rangeOffsetStartFilters: rangeOffsetStartFiltersError,
+      rangeOffsetEndFilters: rangeOffsetEndFiltersError,
     });
 
     return (
-      partitionIdFiltersError === null && customOffsetFiltersError === null
+      partitionIdFiltersError === null &&
+      customOffsetFiltersError === null &&
+      rangeOffsetStartFiltersError === null &&
+      rangeOffsetEndFiltersError === null
     );
   }
 
   function getDefaultOffset(): DefaultOffset {
-    return isDefaultOffset(defaultOffset)
-      ? defaultOffset
-      : initialDefaultOffset;
+    return searchParams.get(NAMES.defaultOffset) as DefaultOffset;
   }
 
   function setDefaultOffset(defaultOffset: DefaultOffset): void {
+    const oldValue = searchParams.get(NAMES.defaultOffset);
     if (!isDefaultOffset(defaultOffset)) {
       searchParams.set(NAMES.defaultOffset, initialDefaultOffset);
     } else {
@@ -95,9 +146,21 @@ function useMessagesFilters(): OffsetFilters {
     setFilterErrors({
       partitionIdFilters: null,
       customOffsetFilters: null,
+      rangeOffsetStartFilters: null,
+      rangeOffsetEndFilters: null,
     });
-    deletePartitionId();
+
+    // If mode is changing from range to custom or vice versa, then no need to delete partition id
+    if (
+      (oldValue !== "custom" && oldValue !== "range") ||
+      (defaultOffset !== "custom" && defaultOffset !== "range")
+    ) {
+      deletePartitionId();
+    }
+
     deleteCustomOffset();
+    deleteRangeOffsetStart();
+    deleteRangeOffsetEnd();
     setSearchParams(searchParams);
   }
 
@@ -133,6 +196,8 @@ function useMessagesFilters(): OffsetFilters {
             : null,
     }));
     searchParams.set(NAMES.customOffset, customOffset);
+    deleteRangeOffsetStart();
+    deleteRangeOffsetEnd();
     setSearchParams(searchParams);
   }
 
@@ -141,12 +206,74 @@ function useMessagesFilters(): OffsetFilters {
     setSearchParams(searchParams);
   }
 
+  function getRangeOffsetStart(): string | null {
+    return searchParams.get(NAMES.rangeOffsetStart);
+  }
+
+  function setRangeOffsetStart(rangeOffsetStart: string): void {
+    if (getDefaultOffset() !== "range") {
+      setDefaultOffset("range");
+    }
+    if (rangeOffsetStart.length === 0) {
+      setFilterErrors((prev) => ({
+        ...prev,
+        rangeOffsetFilters: "Please enter the starting offset",
+      }));
+      deleteRangeOffsetStart();
+      return;
+    }
+    setFilterErrors((prev) => ({
+      ...prev,
+      rangeOffsetFilters:
+        rangeOffsetStart === "" ? "Please enter the starting offset" : null,
+    }));
+    searchParams.set(NAMES.rangeOffsetStart, rangeOffsetStart);
+    deleteCustomOffset();
+    setSearchParams(searchParams);
+  }
+
+  function deleteRangeOffsetStart() {
+    searchParams.delete(NAMES.rangeOffsetStart);
+    setSearchParams(searchParams);
+  }
+
+  function getRangeOffsetEnd(): string | null {
+    return searchParams.get(NAMES.rangeOffsetEnd);
+  }
+
+  function setRangeOffsetEnd(rangeOffsetEnd: string): void {
+    if (getDefaultOffset() !== "range") {
+      setDefaultOffset("range");
+    }
+    if (rangeOffsetEnd.length === 0) {
+      setFilterErrors((prev) => ({
+        ...prev,
+        rangeOffsetFilters: "Please enter the ending offset",
+      }));
+      deleteRangeOffsetEnd();
+      return;
+    }
+    setFilterErrors((prev) => ({
+      ...prev,
+      rangeOffsetFilters:
+        rangeOffsetEnd === "" ? "Please enter the ending offset" : null,
+    }));
+    searchParams.set(NAMES.rangeOffsetEnd, rangeOffsetEnd);
+    deleteCustomOffset();
+    setSearchParams(searchParams);
+  }
+
+  function deleteRangeOffsetEnd() {
+    searchParams.delete(NAMES.rangeOffsetEnd);
+    setSearchParams(searchParams);
+  }
+
   function getPartitionId(): string | null {
     return searchParams.get(NAMES.partitionId);
   }
 
   function setPartitionId(partitionId: string): void {
-    if (getDefaultOffset() !== "custom") {
+    if (getDefaultOffset() !== "custom" && getDefaultOffset() !== "range") {
       setDefaultOffset("custom");
     }
     if (partitionId.length === 0) {
@@ -172,22 +299,28 @@ function useMessagesFilters(): OffsetFilters {
     setSearchParams(searchParams);
   }
 
-  function getFetchingMode() {
-    if (defaultOffset === "custom") {
-      return "Custom";
+  function getFetchingMode(): TopicMessagesFetchModeTypes {
+    if (getDefaultOffset() === "custom") {
+      return "custom";
+    }
+    if (getDefaultOffset() === "range") {
+      return "range";
     }
 
-    return "Default";
+    return "default";
   }
 
   useEffect(() => {
-    if (defaultOffset !== "custom" || defaultOffset === null) {
-      searchParams.set(
-        NAMES.defaultOffset,
-        defaultOffset === null ? initialDefaultOffset : defaultOffset
-      );
-      setSearchParams(searchParams);
-    }
+    const toSetDefaultOffset = !isDefaultOffset(
+      searchParams.get(NAMES.defaultOffset)
+    )
+      ? initialDefaultOffset
+      : searchParams.get(NAMES.defaultOffset);
+    searchParams.set(
+      NAMES.defaultOffset,
+      toSetDefaultOffset === null ? initialDefaultOffset : toSetDefaultOffset
+    );
+    setSearchParams(searchParams);
   }, []);
 
   return {
@@ -204,6 +337,14 @@ function useMessagesFilters(): OffsetFilters {
       setCustomOffset,
       deleteCustomOffset,
     },
+    rangeOffsetFilters: {
+      rangeOffsetStart: getRangeOffsetStart(),
+      setRangeOffsetStart,
+      deleteRangeOffsetStart,
+      rangeOffsetEnd: getRangeOffsetEnd(),
+      setRangeOffsetEnd,
+      deleteRangeOffsetEnd,
+    },
     partitionIdFilters: {
       partitionId: getPartitionId(),
       setPartitionId,
@@ -213,9 +354,10 @@ function useMessagesFilters(): OffsetFilters {
 }
 
 export {
-  useMessagesFilters,
   defaultOffsets,
   isDefaultOffset,
+  useMessagesFilters,
   type DefaultOffset,
   type FilterErrors,
+  type TopicMessagesFetchModeTypes,
 };
